+++ src/equalizer.js
@@ -1,11 +1,12 @@
 import { diffArrays } from 'diff'
+import { parse } from 'intl-messageformat-parser'
 
 import { ERRORS } from './constants'
 import fileReader from './fileReader'
 
 function getExtraKeys(referenceKeys, currentLanguageKeys) {
   return currentLanguageKeys.filter(
-    key => !referenceKeys.some(currKey => currKey === key)
+    (key) => !referenceKeys.some((currKey) => currKey === key)
   )
 }
 
@@ -13,7 +14,8 @@ function getCorrectLines(referenceOrder, languageOrder) {
   const diff = diffArrays(referenceOrder, languageOrder)
 
   const correctLines = diff.reduce((chunks, line) => {
-    const key = line.value[0]
+    const [key] = line.value
+
     if (line.removed || line.added) {
       if (!chunks[key]) {
         chunks[key] = {}
@@ -35,7 +37,7 @@ function getCorrectLines(referenceOrder, languageOrder) {
     return chunks
   }, {})
 
-  return Object.keys(correctLines).map(key => {
+  return Object.keys(correctLines).map((key) => {
     return {
       key,
       ...correctLines[key],
@@ -47,7 +49,7 @@ export function equalize({ languages, localesDirectory, referenceLocale }) {
   const termsPerLanguage = fileReader({ languages, localesDirectory })
 
   const hasEmptyLanguage = languages.some(
-    language =>
+    (language) =>
       !termsPerLanguage[language] ||
       Object.keys(termsPerLanguage[language]).length === 0
   )
@@ -57,7 +59,7 @@ export function equalize({ languages, localesDirectory, referenceLocale }) {
       error: {
         code: ERRORS.ERROR_NO_KEYS_LOCALE,
         data: languages.find(
-          language =>
+          (language) =>
             !termsPerLanguage[language] ||
             Object.keys(termsPerLanguage[language]).length === 0
         ),
@@ -65,24 +67,45 @@ export function equalize({ languages, localesDirectory, referenceLocale }) {
     }
   }
 
+  /** @type {import('./syntaxErrors').EqualizeResult} */
   const processedLanguages = {}
 
-  Object.keys(termsPerLanguage[referenceLocale]).forEach(key => {
-    const referenceOrder = Object.keys(termsPerLanguage[referenceLocale])
+  const referenceOrder = Object.keys(termsPerLanguage[referenceLocale])
 
-    languages.forEach(language => {
+  Object.keys(termsPerLanguage[referenceLocale]).forEach((key) => {
+    languages.forEach((language) => {
       if (!processedLanguages[language]) {
         processedLanguages[language] = {
           missingKeys: [],
           wrongOrderKeys: [],
+          syntaxErrors: [],
+          extraKeys: [],
         }
       }
 
-      if (!termsPerLanguage[language][key]) {
+      const languageTranslation = termsPerLanguage[language][key]
+
+      if (!languageTranslation) {
         processedLanguages[language].missingKeys.push({
           key,
           ...termsPerLanguage[referenceLocale][key],
         })
+      } else {
+        try {
+          // eslint-disable-next-line no-unused-vars
+          const parsedMessage = parse(languageTranslation, {
+            captureLocation: true,
+          })
+
+          // TODO: check if the AST of `parsedMessage` is roughly the same
+          // as the reference locale
+        } catch (err) {
+          processedLanguages[language].syntaxErrors.push({
+            key,
+            error: err,
+            message: languageTranslation,
+          })
+        }
       }
 
       const languageOrder = Object.keys(termsPerLanguage[language])
@@ -112,7 +135,7 @@ export function equalizeRegionLocales({
 
   const processedLanguages = {}
 
-  regionLocales.forEach(region => {
+  regionLocales.forEach((region) => {
     processedLanguages[region] = {
       extraKeys: getExtraKeys(
         Object.keys(termsPerLanguage[referenceLocale]),