+++ src/fileReader.test.js
@@ -5,13 +5,13 @@ jest.mock('fs')
 describe('File Reader', () => {
   it('should return an object by languages and their messages', () => {
     require('fs').__setMockFiles({
-      [`${process.cwd()}/src/__mocks__/locales/en.js`]: '',
-      [`${process.cwd()}/src/__mocks__/locales/pt.json`]: '',
+      [`${process.cwd()}/src/__fixtures__/locales/en.js`]: '',
+      [`${process.cwd()}/src/__fixtures__/locales/pt.json`]: '',
     })
 
     const result = fileReader({
       languages: ['en', 'pt'],
-      localesDirectory: `${process.cwd()}/src/__mocks__/locales`,
+      localesDirectory: `${process.cwd()}/src/__fixtures__/locales`,
     })
 
     expect(result.pt).toBeDefined()
@@ -23,13 +23,13 @@ describe('File Reader', () => {
 
   it('should return an empty object if language file is not found', () => {
     require('fs').__setMockFiles({
-      [`${process.cwd()}/src/__mocks__/locales/en.js`]: '',
-      [`${process.cwd()}/src/__mocks__/locales/pt.json`]: '',
+      [`${process.cwd()}/src/__fixtures__/locales/en.js`]: '',
+      [`${process.cwd()}/src/__fixtures__/locales/pt.json`]: '',
     })
 
     const result = fileReader({
       languages: ['en', 'pt', 'es'],
-      localesDirectory: `${process.cwd()}/src/__mocks__/locales`,
+      localesDirectory: `${process.cwd()}/src/__fixtures__/locales`,
     })
 
     expect(result.pt).toBeDefined()